81: Separate post-related functionality
We want to change our project layout from "package by layer" to "package by layered feature". It's a good practice and will allow us to avoid having huge files with thousands lines of code. You can read more about different project layout approaches [here](https://proandroiddev.com/package-by-type-by-layer-by-feature-vs-package-by-layered-feature-e59921a4dffa) and [here](http://www.javapractices.com/topic/TopicAction.do?Id=205).

As a part of this we need to move all post-related functionality to a separate directory.
From this:
```text
src/
    templates/
    __init__.py
    config.py
    database.py
    forms.py
    models.py
    routes.py
```
To this:
```text
src/
    templates/
    posts/
        templates/
            some-template.html
        __init__.py
        forms.py
        models.py
        routes.py
    __init__.py
    config.py
    database.py
    forms.py
    models.py
    routes.py
```

So in the scope of this task we need to move all post-related code to the separate package inside our `src` directory as described above. And check that everything works fine after these changes.
