+++ src/posts/__init__.py
@@ -1,0 +1,1 @@
+"""'posts' package's initialization module."""
